pyproject.toml
[project]
name = "bitjira-lifter"
version = "0.1.0"
description = "CLI tool for AI-driven commit messages, PR descriptions, and Jira-based branch management"
authors = [ { name="Your Name", email="you@example.com" } ]

[project.dependencies]
click = "^8.0"
typer = {extras=["all"], version="^0.6"}
atlassian-python-api = "^3.9"
openai = "^0.27"
PyYAML = "^6.0" 
[tool.poetry]
name = "bitjira-lifter"
version = "0.1.0"
description = ""
authors = ["Zhima-Mochi <yehgoter@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyperclip = "^1.9.0"
atlassian-python-api = "^4.0.3"
pyyaml = "^6.0.2"
transformers = "^4.51.3"
accelerate = "^1.6.0"


[tool.poetry.group.dev.dependencies]
typer = "^0.15.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
